Move get_byte_field to util.py
